fix issue with unmarshaling response in error case
In cases the request failed, but netcup was able to
send the response, the "responsedata" is actually empty (string),
which in this cases lead to deserialization errors, so a wrong
type of error was propagated to the user. This is now fixed,
tests were added to prevent such behavior in the future.