feat: return a pointer to an empty slice of records
If ``InfoDnsRecords`` or ``UpdateDnsRecords`` fails, the resulting
pointer to slice of DnsRecord points to ``nil``. In some cases though,
the error is recoverable and user wants to proceed, assuming that no
records were returned. In that case, we can help to spare an
additional check by returning a pointer to an empty slice of records.

Fixes: #10